# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API",
        "contact": {
            "email": "se6@rep.earth"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Generation pdf - API server"
        }
    ],
    "paths": {
        "/generate-pdf": {
            "post": {
                "tags": [
                    "WorkFlow"
                ],
                "summary": "Generate pdf by got data and save it in Yandex.Disk (response has link to download and link to review)",
                "operationId": "makePdf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkFlowMakeProcessByDataRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Please, check token access"
                    }
                },
                "security": [
                    {
                        "x-api-token": []
                    }
                ]
            }
        },
        "/delete/{uuid}": {
            "delete": {
                "tags": [
                    "WorkFlow"
                ],
                "summary": "Delete file by uuid from Yandex.Disk and information from database on server",
                "operationId": "deleteById",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "uuid file",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "x-api-token": []
                    }
                ]
            }
        },
        "/delete-all": {
            "delete": {
                "tags": [
                    "WorkFlow"
                ],
                "summary": "Delete all files from Yandex.Disk and information from database on server",
                "operationId": "deleteAll",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "x-api-token": []
                    }
                ]
            }
        },
        "/get-info-disk": {
            "get": {
                "tags": [
                    "YandexDisk"
                ],
                "summary": "Get information about Yandex.Disk",
                "operationId": "infoDisk",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "x-api-token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "WorkFlowMakeProcessByDataRequest": {
                "title": "Example - what data should the body consist of",
                "description": "Some simple request create as example",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Unique ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "data": {
                        "title": "Data",
                        "description": "Name of key for storring",
                        "type": "string",
                        "example": "random"
                    },
                    "value": {
                        "title": "Value",
                        "description": "Value for storring",
                        "type": "string",
                        "example": "awesome"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "x-api-token": {
                "type": "apiKey",
                "name": "x-api-token",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "WorkFlow",
            "description": "Workflow with pdf"
        },
        {
            "name": "YandexDisk",
            "description": "Some jobs with Yandex.Disk"
        }
    ]
}